Basic test infrastructure
I've picked up the idea from @jordimassaguerpla and mocked the official client
through an interface that contains all the required functions for this
application. Then, the `mockClient` struct provides all the needed candy. This
way, we can execute all the tests without even requiring Docker to be there.

This is useful to test which methods are being called and how everything is
setup. Note that this package still needs integration tests: actually running
the command with a docker instance running in the system. This will be done
inside SUSE's infrastructure through Jenkins, but I'd also like to provide some
method to test it locally (e.g. through vagrant).

Moreover, this commit also includes a couple of improvements on the
initialization of the docker client.