Implemented a caching mechanism
We'll be using this for the `images` command. The cache is kept in a JSON file
inside of one of the directories that are commonly used for caching. For now,
we are storing the IDs of both images that are based on SUSE and those that are
not. With this patch, the `images` command goes way faster than before.

I've updated the documentation accordingly.