Workaround special zypper flags
zypper has some special flags that can act both as boolean and string
ones.

For example:
  zypper lp --bugzilla
In the above case --buzilla acts as a boolean flag, while with:
  zypper lp --bugzilla=123
acts like a string flag.
We have to differentiate between invocations with and without the "=".
When the "=" is not found we have to artificially inject an empty string
to avoid the next parameter to be considered the flag value.