Leave early when a signal comes for expensive commands
The `images` and the `ps` commands are expensive since they can potentially
start a huge amount of containers. With this commit, when a signal comes in
the middle of the execution of these commands, said command will leave early.
Since the killing of the possibly generated container has already been done by
the `startContainer` function, and no containers will get created after the
signal gets fired, then we can be sure that the resulting environment will be
clean (we won't leave containers pending to be killed).

Anyways, this is a workaround on the issue #58. This commit does not make
zypper-docker to handle signals perfectly clean, but it's a "good enough"
solution for most cases. I've investigated quite a lot on this, and I've
realized that in order to properly fix this issue, some refactoring have to be
done in the codebase. For safety concerns, I'll do such refactorings after the
first version of zypper-docker is released. For now, this commit will suffice.

Fixes #58

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>