Refactored the test suite and fixed some regressions on logged messages
This is a two-sided commit. On one part, this commit fixes regressions on some
logged messages. In particular, some messages that were actually useful to the
user were always redirected to the log. This was really confusing from a UX
point of view.

Because of the previous point, some tests have to be updated. Moreover, the
other part of this commit was the realization that the test suite was *really*
repetitive, and so a table driven approach would bring some fresh air to the
test suite. Because of that, this commit brings quite a refactoring on the test
suite by transforming (almost) all the commands tests into a table driven
fashion. Moreover, I've added more test cases here and there, so the test suite
now covers a bit more of code.

Finally, this commit also fixes a regression in the integration tests. In
particular: 6aee4f4722d8706985af705b14ab104212ae2360. To note: codegangsta/cli
does not play well with `Before` hooks for each command.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>