Ignore certain exit codes of zypper
Do not cause a patch/update operation to fail just because zypper
didn't exit with 0. In certain cases we can ignore an exit code that is
different from 0.

For example: if zypper updates itself it will recommend to restart the
system and will exit with 103. In that case we can safely ignore the
exit code and commit the changes to a new image.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>