climate: get coverage output
Don't pipe scripts/climate to /dev/null to let the user know why climate
fails. Also add a new option to climate (-p) to preserve the coverage
profile and not delete it, so that it can be examined by the user.
Notice that the coverage profile has been renamed from 'c.out' to
'zypper-docker.cov' to be more expressive.

Signed-off-by: Valentin Rothberg <vrothberg@suse.com>