Fix a bug that caused the force flag to not work
The function that resets the cache if the force flag is
used did not work as intended.

Previously the reset() method which is called if the force
flag is set, made use of the flush() method, which compares
the existing cache file with the changes made, appends the
data to the existing file and removes redundant data.
If the changes made to the file are empty strings as in
case of the reset() method and the flush method is called
this results in the cache file not to be changed at all.

To make sure the reset() method works as intended the
cache file is overwritten with empty strings directly
instead of calling the flush() method.

Signed-off-by: Pascal Arlt <parlt@suse.com>