Fix the docker ps command
Previously the docker psCmd function made use of getImageID
which expects a repo:tag combination, but was called with the
ID which the function was expected to return. Thus the call
of getImageID was redundant in this place.
The tests were updated to reflect the new behaviour.

Another problem that influenced zypper-docker ps was that
an image was not added to the outdated string in the cache
file after a zypper-docker up or patch if the image to be
updated was parsed with it's image ID.

Fixes: #144
Fixes: #145
Signed-off-by: Pascal Arlt <parlt@suse.com>