Analyze container instead of base image by default
Committing running containers to a new image before analyzing
them with the zypper-docker luc, lpc and pchkc commands is now
the default. This way changes made to the running container will
always be included.
If the base image of a container should be analyzed, which was
the former default, a new --base flag can be used.
e.g. zypper-docker pchkc --base <container id>

Signed-off-by: Pascal Arlt <parlt@suse.com>