Use custom cache dir for zypper up/patch commands
Previously the default zypper cache directory was used for
update/patch commands. This cache dir was emptied afterwards.
If a user already had data from previous zypper commands in the
cache, the data would be lost after a zypper-docker up/patch.

Now a custom directory is used for the zypper-docker commands
that will be deleted before the updated image is created.
This way the default zypper cache dir will be preserved.

Signed-off-by: Pascal Arlt <parlt@suse.com>